+++ src/consensus/aft/raft.h
@@ -381,7 +381,7 @@ namespace aft
 
       std::lock_guard<std::mutex> guard(state->lock);
       state->current_view += starting_view_change;
-      become_leader();
+      become_leader(true);
     }
 
     void force_become_leader(
@@ -405,7 +405,7 @@ namespace aft
       state->view_history.initialise(terms);
       state->view_history.update(index, term);
       state->current_view += starting_view_change;
-      become_leader();
+      become_leader(true);
     }
 
     void init_as_follower(
@@ -434,7 +434,7 @@ namespace aft
     Index get_commit_idx()
     {
       std::lock_guard<std::mutex> guard(state->lock);
-      return state->commit_idx;
+      return get_commit_idx_unsafe();
     }
 
     Term get_term()
@@ -446,14 +446,17 @@ namespace aft
     std::pair<Term, Index> get_commit_term_and_idx()
     {
       std::lock_guard<std::mutex> guard(state->lock);
-      return {get_term_internal(state->commit_idx), state->commit_idx};
+      ccf::SeqNo commit_idx = get_commit_idx_unsafe();
+      return {get_term_internal(commit_idx), commit_idx};
     }
 
     std::optional<kv::Consensus::SignableTxIndices>
     get_signable_commit_term_and_idx()
     {
       std::lock_guard<std::mutex> guard(state->lock);
-      if (state->commit_idx >= election_index)
+      if (
+        consensus_type == ConsensusType::BFT ||
+        state->commit_idx >= election_index)
       {
         kv::Consensus::SignableTxIndices r;
         r.term = get_term_internal(state->commit_idx);
@@ -955,7 +958,10 @@ namespace aft
           vc->serialize(data, size);
           CCF_ASSERT_FMT(size == 0, "Did not write everything");
 
-          LOG_INFO_FMT("Sending view change msg view:{}", vcm.view);
+          LOG_INFO_FMT(
+            "Sending view change msg view:{}, primary_at_view:{}",
+            vcm.view,
+            get_primary(vcm.view));
           for (auto it = nodes.begin(); it != nodes.end(); ++it)
           {
             auto to = it->first;
@@ -1016,6 +1022,7 @@ namespace aft
       const uint8_t* data,
       size_t size)
     {
+      LOG_DEBUG_FMT("Received evidence for view:{}, from:{}", r.view, from);
       auto node = nodes.find(from);
       if (node == nodes.end())
       {
@@ -1167,6 +1174,8 @@ namespace aft
 
     void append_new_view(ccf::View view)
     {
+      LOG_INFO_FMT(
+        "Writing view change to ledger as a new primay, view:{}", view);
       state->current_view = view;
       become_leader();
       state->new_view_idx =
@@ -1237,6 +1246,19 @@ namespace aft
         return ccf::VIEW_UNKNOWN;
 
       return state->view_history.view_at(idx);
+    }
+
+    Index get_commit_idx_unsafe()
+    {
+      if (consensus_type == ConsensusType::CFT)
+      {
+        return state->commit_idx;
+      }
+      else
+      {
+        auto progress_tracker = store->get_progress_tracker();
+        return progress_tracker->get_highest_committed_level();
+      }
     }
 
     void send_append_entries(const ccf::NodeId& to, Index start_idx)
@@ -1529,7 +1551,13 @@ namespace aft
             "rolling back from {} to {}",
             state->last_idx,
             r.prev_idx);
-          rollback(r.prev_idx);
+          auto rollback_level = r.prev_idx;
+          if (consensus_type == ConsensusType::BFT)
+          {
+            auto progress_tracker = store->get_progress_tracker();
+            rollback_level = progress_tracker->get_rollback_seqno();
+          }
+          rollback(rollback_level);
         }
         else
         {
@@ -2282,7 +2310,8 @@ namespace aft
             nonce.value(),
             state->my_node_id,
             get_last_configuration_nodes(),
-            is_primary());
+            is_primary(),
+            tx_id.seqno <= state->last_idx);
           break;
         }
         default:
@@ -2317,7 +2346,8 @@ namespace aft
         r.nonce,
         from,
         get_last_configuration_nodes(),
-        is_primary());
+        is_primary(),
+        r.idx <= state->last_idx);
 
       update_commit();
     }
@@ -2405,6 +2435,7 @@ namespace aft
           reinterpret_cast<uint8_t*>(&vw),
           reinterpret_cast<uint8_t*>(&vw) + sizeof(ViewChangeEvidenceMsg));
 
+        LOG_DEBUG_FMT("Sending evidence to:{}", from);
         channels->send_authenticated(
           from, ccf::NodeMsgType::consensus_msg, data);
       }
@@ -2639,14 +2670,26 @@ namespace aft
       }
     }
 
-    void become_leader()
+    void become_leader(bool force_become_leader = false)
     {
       if (is_retired())
       {
         return;
       }
 
-      election_index = last_committable_index();
+      // When we force to become the primary we are going around the
+      // consensus protocol. This only happens when a node starts a new network
+      // and have a gensis or recovery tx as the last transaction, for BFT this
+      // transaction is not prepared and but must not be rolled back.
+      if (consensus_type == ConsensusType::BFT && !force_become_leader)
+      {
+        auto progress_tracker = store->get_progress_tracker();
+        election_index = progress_tracker->get_rollback_seqno();
+      }
+      else
+      {
+        election_index = last_committable_index();
+      }
       LOG_DEBUG_FMT(
         "Election index is {} in term {}", election_index, state->current_view);
       // Discard any un-committable updates we may hold,
@@ -2807,7 +2850,7 @@ namespace aft
       auto progress_tracker = store->get_progress_tracker();
       if (progress_tracker != nullptr)
       {
-        new_commit_bft_idx = progress_tracker->get_highest_committed_nonce();
+        new_commit_bft_idx = progress_tracker->get_highest_committed_level();
       }
 
       // Obtain CFT watermarks
@@ -2886,7 +2929,17 @@ namespace aft
         }
 
         if (can_commit)
-          commit(highest_committable);
+        {
+          if (consensus_type == ConsensusType::CFT)
+          {
+            commit(highest_committable);
+          }
+          else
+          {
+            auto progress_tracker = store->get_progress_tracker();
+            commit(progress_tracker->get_highest_committed_level());
+          }
+        }
       }
     }
 
@@ -3053,13 +3106,17 @@ namespace aft
 
     void rollback(Index idx)
     {
-      if (idx < state->commit_idx)
+      if (
+        (consensus_type == ConsensusType::CFT && idx < state->commit_idx) ||
+        (consensus_type == ConsensusType::BFT &&
+         idx < state->bft_watermark_idx))
       {
         LOG_FAIL_FMT(
-          "Asked to rollback to {} but committed to {} - ignoring rollback "
-          "request",
+          "Asked to rollback to idx:{} but committed to commit_idx:{}, "
+          "bft_watermark_idx:{} - ignoring rollback request",
           idx,
-          state->commit_idx);
+          state->commit_idx,
+          state->bft_watermark_idx);
         return;
       }
 
@@ -3116,12 +3173,6 @@ namespace aft
       if (changed)
       {
         create_and_remove_node_state();
-      }
-
-      if (consensus_type == ConsensusType::BFT)
-      {
-        auto progress_tracker = store->get_progress_tracker();
-        progress_tracker->rollback(idx, state->current_view);
       }
     }